+++ pkg/kernel/lib/transformations/continuation.dart
@@ -12,29 +12,34 @@ import '../visitor.dart';
 
 import 'async.dart';
 
-void transformLibraries(CoreTypes coreTypes, List<Library> libraries) {
+void transformLibraries(
+    CoreTypes coreTypes, List<Library> libraries, bool syncAsync) {
   var helper = new HelperNodes.fromCoreTypes(coreTypes);
-  var rewriter = new RecursiveContinuationRewriter(helper);
+  var rewriter = new RecursiveContinuationRewriter(helper, syncAsync);
   for (var library in libraries) {
     rewriter.rewriteLibrary(library);
   }
 }
 
-Program transformProgram(CoreTypes coreTypes, Program program) {
+Program transformProgram(CoreTypes coreTypes, Program program, bool syncAsync) {
   var helper = new HelperNodes.fromCoreTypes(coreTypes);
-  var rewriter = new RecursiveContinuationRewriter(helper);
+  var rewriter = new RecursiveContinuationRewriter(helper, syncAsync);
   return rewriter.rewriteProgram(program);
 }
 
 class RecursiveContinuationRewriter extends Transformer {
   final HelperNodes helper;
+
+  /// Whether `async` functions should start synchronously.
+  final bool syncAsync;
+
   final VariableDeclaration asyncJumpVariable = new VariableDeclaration(
       ":await_jump_var",
       initializer: new IntLiteral(0));
   final VariableDeclaration asyncContextVariable =
       new VariableDeclaration(":await_ctx_var");
 
-  RecursiveContinuationRewriter(this.helper);
+  RecursiveContinuationRewriter(this.helper, this.syncAsync);
 
   Program rewriteProgram(Program node) {
     return node.accept(this);
@@ -52,14 +57,15 @@ class RecursiveContinuationRewriter extends Transformer {
     switch (node.asyncMarker) {
       case AsyncMarker.Sync:
       case AsyncMarker.SyncYielding:
-        node.transformChildren(new RecursiveContinuationRewriter(helper));
+        node.transformChildren(
+            new RecursiveContinuationRewriter(helper, syncAsync));
         return node;
       case AsyncMarker.SyncStar:
-        return new SyncStarFunctionRewriter(helper, node).rewrite();
+        return new SyncStarFunctionRewriter(helper, node, syncAsync).rewrite();
       case AsyncMarker.Async:
-        return new AsyncFunctionRewriter(helper, node).rewrite();
+        return new AsyncFunctionRewriter(helper, node, syncAsync).rewrite();
       case AsyncMarker.AsyncStar:
-        return new AsyncStarFunctionRewriter(helper, node).rewrite();
+        return new AsyncStarFunctionRewriter(helper, node, syncAsync).rewrite();
     }
   }
 }
@@ -72,8 +78,9 @@ abstract class ContinuationRewriterBase extends RecursiveContinuationRewriter {
   int capturedTryDepth = 0; // Deepest yield point within a try-block.
   int capturedCatchDepth = 0; // Deepest yield point within a catch-block.
 
-  ContinuationRewriterBase(HelperNodes helper, this.enclosingFunction)
-      : super(helper);
+  ContinuationRewriterBase(
+      HelperNodes helper, this.enclosingFunction, bool syncAsync)
+      : super(helper, syncAsync);
 
   /// Given a container [type], which is an instantiation of the given
   /// [containerClass] extract its element type.
@@ -157,13 +164,14 @@ abstract class ContinuationRewriterBase extends RecursiveContinuationRewriter {
 class SyncStarFunctionRewriter extends ContinuationRewriterBase {
   final VariableDeclaration iteratorVariable;
 
-  SyncStarFunctionRewriter(HelperNodes helper, FunctionNode enclosingFunction)
+  SyncStarFunctionRewriter(
+      HelperNodes helper, FunctionNode enclosingFunction, syncAsync)
       : iteratorVariable = new VariableDeclaration(':iterator')
           ..type = new InterfaceType(helper.syncIteratorClass, [
             ContinuationRewriterBase.elementTypeFrom(
                 helper.iterableClass, enclosingFunction.returnType)
           ]),
-        super(helper, enclosingFunction);
+        super(helper, enclosingFunction, syncAsync);
 
   FunctionNode rewrite() {
     // :sync_op(:iterator) {
@@ -256,8 +264,9 @@ abstract class AsyncRewriterBase extends ContinuationRewriterBase {
 
   ExpressionLifter expressionRewriter;
 
-  AsyncRewriterBase(helper, enclosingFunction)
-      : super(helper, enclosingFunction) {}
+  AsyncRewriterBase(
+      HelperNodes helper, FunctionNode enclosingFunction, bool syncAsync)
+      : super(helper, enclosingFunction, syncAsync) {}
 
   void setupAsyncContinuations(List<Statement> statements) {
     expressionRewriter = new ExpressionLifter(this);
@@ -754,8 +763,9 @@ abstract class AsyncRewriterBase extends ContinuationRewriterBase {
 class AsyncStarFunctionRewriter extends AsyncRewriterBase {
   VariableDeclaration controllerVariable;
 
-  AsyncStarFunctionRewriter(helper, enclosingFunction)
-      : super(helper, enclosingFunction);
+  AsyncStarFunctionRewriter(
+      HelperNodes helper, FunctionNode enclosingFunction, bool syncAsync)
+      : super(helper, enclosingFunction, syncAsync);
 
   FunctionNode rewrite() {
     var statements = <Statement>[];
@@ -862,8 +872,9 @@ class AsyncFunctionRewriter extends AsyncRewriterBase {
   VariableDeclaration completerVariable;
   VariableDeclaration returnVariable;
 
-  AsyncFunctionRewriter(helper, enclosingFunction)
-      : super(helper, enclosingFunction);
+  AsyncFunctionRewriter(
+      HelperNodes helper, FunctionNode enclosingFunction, bool syncAsync)
+      : super(helper, enclosingFunction, syncAsync);
 
   FunctionNode rewrite() {
     var statements = <Statement>[];
@@ -877,16 +888,29 @@ class AsyncFunctionRewriter extends AsyncRewriterBase {
     final DartType returnType =
         new InterfaceType(helper.futureOrClass, <DartType>[valueType]);
     var completerTypeArguments = <DartType>[valueType];
-    var completerType =
-        new InterfaceType(helper.completerClass, completerTypeArguments);
 
-    // final Completer<T> :completer = new Completer<T>.sync();
-    completerVariable = new VariableDeclaration(":completer",
-        initializer: new StaticInvocation(helper.completerConstructor,
-            new Arguments([], types: completerTypeArguments))
-          ..fileOffset = enclosingFunction.body?.fileOffset ?? -1,
-        isFinal: true,
-        type: completerType);
+    if (syncAsync) {
+      final completerType = new InterfaceType(
+          helper.asyncAwaitCompleterClass, completerTypeArguments);
+      // final Completer<T> :completer = new _AsyncAwaitCompleter<T>();
+      completerVariable = new VariableDeclaration(":completer",
+          initializer: new ConstructorInvocation(
+              helper.asyncAwaitCompleterConstructor,
+              new Arguments([], types: completerTypeArguments))
+            ..fileOffset = enclosingFunction.body?.fileOffset ?? -1,
+          isFinal: true,
+          type: completerType);
+    } else {
+      final completerType =
+          new InterfaceType(helper.completerClass, completerTypeArguments);
+      // final Completer<T> :completer = new Completer<T>.sync();
+      completerVariable = new VariableDeclaration(":completer",
+          initializer: new StaticInvocation(helper.completerConstructor,
+              new Arguments([], types: completerTypeArguments))
+            ..fileOffset = enclosingFunction.body?.fileOffset ?? -1,
+          isFinal: true,
+          type: completerType);
+    }
     statements.add(completerVariable);
 
     returnVariable = new VariableDeclaration(":return_value", type: returnType);
@@ -894,14 +918,23 @@ class AsyncFunctionRewriter extends AsyncRewriterBase {
 
     setupAsyncContinuations(statements);
 
-    // new Future.microtask(:async_op);
-    var newMicrotaskStatement = new ExpressionStatement(new StaticInvocation(
-        helper.futureMicrotaskConstructor,
-        new Arguments([new VariableGet(nestedClosureVariable)],
-            types: [const DynamicType()]))
-      ..fileOffset = enclosingFunction.fileOffset);
-    statements.add(newMicrotaskStatement);
-
+    if (syncAsync) {
+      // :completer.start(:async_op);
+      var startStatement = new ExpressionStatement(new MethodInvocation(
+          new VariableGet(completerVariable),
+          new Name('start'),
+          new Arguments([new VariableGet(nestedClosureVariable)]))
+        ..fileOffset = enclosingFunction.fileOffset);
+      statements.add(startStatement);
+    } else {
+      // new Future.microtask(:async_op);
+      var newMicrotaskStatement = new ExpressionStatement(new StaticInvocation(
+          helper.futureMicrotaskConstructor,
+          new Arguments([new VariableGet(nestedClosureVariable)],
+              types: [const DynamicType()]))
+        ..fileOffset = enclosingFunction.fileOffset);
+      statements.add(newMicrotaskStatement);
+    }
     // return :completer.future;
     var completerGet = new VariableGet(completerVariable);
     var returnStatement = new ReturnStatement(new PropertyGet(completerGet,
@@ -964,9 +997,11 @@ class HelperNodes {
   final Procedure asyncThenWrapper;
   final Procedure awaitHelper;
   final Class completerClass;
+  final Class asyncAwaitCompleterClass;
   final Member completerComplete;
   final Member completerCompleteError;
   final Member completerConstructor;
+  final Member asyncAwaitCompleterConstructor;
   final Member completerFuture;
   final Library coreLibrary;
   final CoreTypes coreTypes;
@@ -1001,9 +1036,11 @@ class HelperNodes {
       this.asyncThenWrapper,
       this.awaitHelper,
       this.completerClass,
+      this.asyncAwaitCompleterClass,
       this.completerComplete,
       this.completerCompleteError,
       this.completerConstructor,
+      this.asyncAwaitCompleterConstructor,
       this.completerFuture,
       this.coreLibrary,
       this.coreTypes,
@@ -1039,9 +1076,11 @@ class HelperNodes {
         coreTypes.asyncThenWrapperHelperProcedure,
         coreTypes.awaitHelperProcedure,
         coreTypes.completerClass,
+        coreTypes.asyncAwaitCompleterClass,
         coreTypes.completerComplete,
         coreTypes.completerCompleteError,
         coreTypes.completerSyncConstructor,
+        coreTypes.asyncAwaitCompleterConstructor,
         coreTypes.completerFuture,
         coreTypes.coreLibrary,
         coreTypes,