+++ pkg/vm/bin/kernel_service.dart
@@ -48,7 +48,9 @@ abstract class Compiler {
   CompilerOptions options;
 
   Compiler(this.fileSystem, Uri platformKernel,
-      {this.strongMode: false, bool suppressWarnings: false}) {
+      {this.strongMode: false,
+      bool suppressWarnings: false,
+      bool syncAsync: false}) {
     Uri packagesUri = (Platform.packageConfig != null)
         ? Uri.parse(Platform.packageConfig)
         : null;
@@ -59,12 +61,14 @@ abstract class Compiler {
       print("DFE: Platform.resolvedExecutable: ${Platform.resolvedExecutable}");
       print("DFE: platformKernel: ${platformKernel}");
       print("DFE: strongMode: ${strongMode}");
+      print("DFE: syncAsync: ${syncAsync}");
     }
 
     options = new CompilerOptions()
       ..strongMode = strongMode
       ..fileSystem = fileSystem
-      ..target = new VmTarget(new TargetFlags(strongMode: strongMode))
+      ..target = new VmTarget(
+          new TargetFlags(strongMode: strongMode, syncAsync: syncAsync))
       ..packagesFileUri = packagesUri
       ..sdkSummary = platformKernel
       ..verbose = verbose
@@ -99,9 +103,11 @@ class IncrementalCompiler extends Compiler {
   IncrementalKernelGenerator generator;
 
   IncrementalCompiler(FileSystem fileSystem, Uri platformKernel,
-      {bool strongMode: false, bool suppressWarnings: false})
+      {bool strongMode: false, bool suppressWarnings: false, syncAsync: false})
       : super(fileSystem, platformKernel,
-            strongMode: strongMode, suppressWarnings: suppressWarnings);
+            strongMode: strongMode,
+            suppressWarnings: suppressWarnings,
+            syncAsync: syncAsync);
 
   @override
   Future<Program> compileInternal(Uri script) async {
@@ -122,9 +128,12 @@ class SingleShotCompiler extends Compiler {
   SingleShotCompiler(FileSystem fileSystem, Uri platformKernel,
       {this.requireMain: false,
       bool strongMode: false,
-      bool suppressWarnings: false})
+      bool suppressWarnings: false,
+      bool syncAsync: false})
       : super(fileSystem, platformKernel,
-            strongMode: strongMode, suppressWarnings: suppressWarnings);
+            strongMode: strongMode,
+            suppressWarnings: suppressWarnings,
+            syncAsync: syncAsync);
 
   @override
   Future<Program> compileInternal(Uri script) async {
@@ -138,7 +147,9 @@ final Map<int, Compiler> isolateCompilers = new Map<int, Compiler>();
 
 Future<Compiler> lookupOrBuildNewIncrementalCompiler(
     int isolateId, List sourceFiles, Uri platformKernel,
-    {bool strongMode: false, bool suppressWarnings: false}) async {
+    {bool strongMode: false,
+    bool suppressWarnings: false,
+    bool syncAsync: false}) async {
   IncrementalCompiler compiler;
   if (isolateCompilers.containsKey(isolateId)) {
     compiler = isolateCompilers[isolateId];
@@ -162,7 +173,9 @@ Future<Compiler> lookupOrBuildNewIncrementalCompiler(
     // isolate needs to receive a message indicating that particular
     // isolate was shut down. Message should be handled here in this script.
     compiler = new IncrementalCompiler(fileSystem, platformKernel,
-        strongMode: strongMode, suppressWarnings: suppressWarnings);
+        strongMode: strongMode,
+        suppressWarnings: suppressWarnings,
+        syncAsync: syncAsync);
     isolateCompilers[isolateId] = compiler;
   }
   return compiler;
@@ -189,6 +202,7 @@ Future _processLoadRequest(request) async {
   final int isolateId = request[6];
   final List sourceFiles = request[7];
   final bool suppressWarnings = request[8];
+  final bool syncAsync = request[9];
 
   Compiler compiler;
   // TODO(aam): There should be no need to have an option to choose
@@ -198,7 +212,7 @@ Future _processLoadRequest(request) async {
   if (incremental) {
     compiler = await lookupOrBuildNewIncrementalCompiler(
         isolateId, sourceFiles, platformKernel,
-        suppressWarnings: suppressWarnings);
+        suppressWarnings: suppressWarnings, syncAsync: syncAsync);
   } else {
     final FileSystem fileSystem = sourceFiles == null
         ? StandardFileSystem.instance
@@ -206,7 +220,8 @@ Future _processLoadRequest(request) async {
     compiler = new SingleShotCompiler(fileSystem, platformKernel,
         requireMain: sourceFiles == null,
         strongMode: strong,
-        suppressWarnings: suppressWarnings);
+        suppressWarnings: suppressWarnings,
+        syncAsync: syncAsync);
   }
 
   CompilationResult result;
@@ -293,6 +308,7 @@ train(String scriptUri, String platformKernel) {
     1 /* isolateId chosen randomly */,
     null /* source files */,
     false /* suppress warnings */,
+    false /* synchronous async */,
   ];
   _processLoadRequest(request);
 }